requirements.txt
python-telegram-bot==20.4
pyrogram==2.0.106
tgcrypto==1.2.5
groq==0.4.0
python-dotenv==1.0.0
requests==2.31.0 